First chunk of compiler work; splitting DCPU-16 kernel
With the right definitions of `CODE` and `END-CODE`, the kernel words
that are too long to be inlined can be reused between the threaded
DCPU-16 kernel and the native code compiled one.

Shorter kernel words can be compiled by inlining.
Prepared for testing a basic app. The tests don't work since the system
isn't self-hosting.